[src/get_si.c] forgot cast when long > limb
[From SVN r13115 (trunk)]

+++ src/get_si.c
@@ -88,7 +88,7 @@ mpfr_get_si (mpfr_srcptr f, mpfr_rnd_t rnd)
         if (exp <= GMP_NUMB_BITS)
           s += MPFR_MANT(x)[n - 1] >> (GMP_NUMB_BITS - exp);
         else
-          s += MPFR_MANT(x)[n - 1] << (exp - GMP_NUMB_BITS);
+          s += (long) MPFR_MANT(x)[n - 1] << (exp - GMP_NUMB_BITS);
         n --;
         exp -= GMP_NUMB_BITS;
       }
